pyproject.toml
[project]
name = "py-nbp-api"
version = "1.0.0"
description = """
Tidy downloading of National Bank of Poland interest rates,
FX rates and gold price data
"""
readme = "README.md"
requires-python = ">=3.11"
keywords = [
    "NBP", "FX", "PLN", "gold", "API", "REST-API", "FX-rates", "zloty"
]
authors = [
  {name = "Artur Wegrzyn", email = "awegrzyn17@gmail.com" }
]
maintainers = [
  {name = "Artur Wegrzyn", email = "awegrzyn17@gmail.com" }
]
dependencies = [
    "pandas",
    "requests",
    "python-dateutil",
    "loguru"
]

[build-system]
requires = ["setuptools>=69.0.3", "wheel"]
build-backend = "setuptools.build_meta"